Admin: split into functional mixin classes (_acls, _configs, _groups, _metadata, _records, _topics)
