+++ kafka/admin/__init__.py
@@ -1,9 +1,8 @@
-from kafka.admin.config_resource import ConfigResource, ConfigResourceType
+from kafka.admin._configs import ConfigResource, ConfigResourceType
 from kafka.admin.client import KafkaAdminClient
-from kafka.admin.acl_resource import (ACL, ACLFilter, ResourcePattern, ResourcePatternFilter, ACLOperation,
-                                      ResourceType, ACLPermissionType, ACLResourcePatternType)
-from kafka.admin.new_topic import NewTopic
-from kafka.admin.new_partitions import NewPartitions
+from kafka.admin._acls import (ACL, ACLFilter, ResourcePattern, ResourcePatternFilter, ACLOperation,
+                                ResourceType, ACLPermissionType, ACLResourcePatternType)
+from kafka.admin._topics import NewTopic, NewPartitions
 
 __all__ = [
     'ConfigResource', 'ConfigResourceType', 'KafkaAdminClient', 'NewTopic', 'NewPartitions', 'ACL', 'ACLFilter',
