+++ kafka/admin/client.py
@@ -1,40 +1,36 @@
-from collections import defaultdict
+"""KafkaAdminClient — high-level Kafka cluster administration."""
+
 import copy
-import itertools
 import logging
 import selectors
 import socket
 import time
-import uuid
 
-from . import ConfigResourceType
-
-from kafka.admin.acl_resource import ACLOperation, ACLPermissionType, ACLFilter, ACL, ResourcePattern, ResourceType, \
-    ACLResourcePatternType, valid_acl_operations
-from kafka.net.compat import KafkaNetClient
-from kafka.protocol.consumer.metadata import (
-    ConsumerProtocolSubscription, ConsumerProtocolAssignment, ConsumerProtocolType,
-)
 import kafka.errors as Errors
-from kafka.errors import (
-    IncompatibleBrokerVersion, KafkaConfigurationError, UnknownTopicOrPartitionError,
-    UnrecognizedBrokerVersion, IllegalArgumentError)
-from kafka.future import Future
+from kafka.errors import KafkaConfigurationError, UnrecognizedBrokerVersion
 from kafka.metrics import MetricConfig, Metrics
-from kafka.protocol.admin import (
-    CreateTopicsRequest, DeleteTopicsRequest, DescribeConfigsRequest, AlterConfigsRequest, CreatePartitionsRequest,
-    ListGroupsRequest, DescribeGroupsRequest, DescribeAclsRequest, CreateAclsRequest, DeleteAclsRequest,
-    DeleteGroupsRequest, DeleteRecordsRequest, DescribeLogDirsRequest, ElectLeadersRequest, ElectionType)
-from kafka.protocol.consumer import OffsetFetchRequest
+from kafka.net.compat import KafkaNetClient
 from kafka.protocol.metadata import MetadataRequest, FindCoordinatorRequest
-from kafka.structs import TopicPartition, OffsetAndMetadata, MemberInformation, GroupInformation
 from kafka.version import __version__
 
+from kafka.admin._acls import ACLAdminMixin
+from kafka.admin._configs import ConfigAdminMixin
+from kafka.admin._groups import GroupAdminMixin
+from kafka.admin._metadata import MetadataAdminMixin
+from kafka.admin._records import RecordAdminMixin
+from kafka.admin._topics import TopicAdminMixin
 
 log = logging.getLogger(__name__)
 
 
-class KafkaAdminClient:
+class KafkaAdminClient(
+    TopicAdminMixin,
+    MetadataAdminMixin,
+    ACLAdminMixin,
+    ConfigAdminMixin,
+    GroupAdminMixin,
+    RecordAdminMixin,
+):
     """A class for administering the Kafka cluster.
 
     Warning:
@@ -240,15 +236,10 @@ class KafkaAdminClient:
         log.debug("KafkaAdminClient is now closed.")
 
     def _validate_timeout(self, timeout_ms):
-        """Validate the timeout is set or use the configuration default.
-
-        Arguments:
-            timeout_ms: The timeout provided by api call, in milliseconds.
-
-        Returns:
-            The timeout to use for the operation.
-        """
+        """Validate the timeout is set or use the configuration default."""
         return timeout_ms or self.config['request_timeout_ms']
+
+    # -- Routing primitives (used by mixins) ----------------------------------
 
     async def _refresh_controller_id(self, timeout_ms=30000):
         """Determine the Kafka cluster controller."""
@@ -274,15 +265,6 @@ class KafkaAdminClient:
 
         Results are cached; subsequent calls for the same group_id return
         the cached value without a network round-trip.
-
-        Arguments:
-            group_id (str): The consumer group ID.
-
-        Returns:
-            int: The node_id of the group's coordinator broker.
-
-        Raises:
-            Errors from the FindCoordinatorResponse (e.g., CoordinatorNotAvailableError).
         """
         cached = self._coordinator_cache.get(group_id)
         if cached is not None:
@@ -300,20 +282,8 @@ class KafkaAdminClient:
     async def _send_request_to_controller(self, request, get_errors_fn=lambda r: (), raise_errors=True, ignore_errors=()):
         """Send a Kafka protocol message to the cluster controller.
 
-        Will block until the message result is received.
-
-        Arguments:
-            request: The message to send.
-
-        Keyword Arguments:
-            get_errors_fn (func): Function to process response and return an iterable of Error types.
-            ignore_errors (tuple): Any non-zero error codes that should be ignored. Not used if raise_errors=False.
-            raise_errors (bool): Whether to raise unhandled errors (True, default) or return response with errors (False).
-
-        Returns:
-            The Kafka protocol response for the message.
+        Retries once on NotControllerError after refreshing the controller id.
         """
-        # retry in case our controller_id is out of date
         if self._controller_id is None or self._controller_id == -1:
             self._controller_id = await self._refresh_controller_id()
 
@@ -334,1279 +304,3 @@ class KafkaAdminClient:
                     "Request '{}' failed with response '{}'."
                     .format(request, response))
         return response
-
-    def create_topics(self, new_topics, timeout_ms=None, validate_only=False, raise_errors=True,
-                      wait_for_metadata=False):
-        """Create new topics in the cluster.
-
-        Arguments:
-            new_topics: A list of topic names,
-                or a dict of {topic_name: {num_partitions:, replication_factor:,
-                                           assignments: {partition: [broker_ids]},
-                                           configs: {key: value}}}
-                List of NewTopic objects is deprecated.
-
-        Keyword Arguments:
-            timeout_ms (numeric, optional): Milliseconds to wait for new topics to be created
-                before the broker returns.
-            validate_only (bool, optional): If True, don't actually create new topics.
-                Not supported by all versions. Default: False
-            raise_errors (bool, optional): Whether to raise errors as exceptions. Default True.
-            wait_for_metadata (bool, optional): If True, after the broker successfully
-                accepts the create request, block until each new topic is visible in
-                broker metadata with a leader assigned for every partition. Useful on
-                KRaft clusters, where CreateTopicsResponse returning NoError does not
-                guarantee that a subsequent MetadataRequest will see the topic. Has no
-                effect when ``validate_only`` is True. Uses a fixed 10-second timeout;
-                call :meth:`wait_for_topics` directly for finer control.
-                Mutually exclusive with validate_only. Default: False
-
-        Returns: CreateTopicResponse
-        """
-        if validate_only and wait_for_metadata:
-            raise ValueError('validate_only and wait_for_metadata are mutually exclusive')
-        timeout_ms = self._validate_timeout(timeout_ms)
-        if validate_only and self._manager.broker_version < (0, 10, 2):
-            raise IncompatibleBrokerVersion(
-                "validate_only requires CreateTopicsRequest >= v1, which is not supported by Kafka {}."
-                .format(self._manager.broker_version))
-
-        _Topic = CreateTopicsRequest.CreatableTopic
-        _Assignment = _Topic.CreatableReplicaAssignment
-        _Config = _Topic.CreatableTopicConfig
-        topics = []
-        if isinstance(new_topics, dict):
-            # {topic_name: {num_partitions:, replication_factor:, assignments: {partition: [broker_ids]}, configs: {key: value}}
-            for topic, data in new_topics.items():
-                configs = data.get('configs', {})
-                topics.append(_Topic(
-                    name=topic,
-                    num_partitions=data.get('num_partitions', -1),
-                    replication_factor=data.get('replication_factor', -1),
-                    assignments=[_Assignment(partition_index=partition_id, broker_ids=replicas)
-                                 for partition_id, replicas in data.get('assignments', {}).items()],
-                    configs=[_Config(name=config_key, value=config_value)
-                             for config_key, config_value in data.get('configs', {}).items()]
-                ))
-        elif all(isinstance(v, str) for v in new_topics):
-            for new_topic in new_topics:
-                topics.append(_Topic(name=new_topic))
-        else:
-            from .new_topic import NewTopic
-            if all(isinstance(v, NewTopic) for v in new_topics):
-                for new_topic in new_topics:
-                    topics.append(_Topic(
-                        name=new_topic.name,
-                        num_partitions=new_topic.num_partitions,
-                        replication_factor=new_topic.replication_factor,
-                        assignments=[_Assignment(partition_index=partition_id, broker_ids=replicas)
-                                     for partition_id, replicas in new_topic.replica_assignments.items()],
-                        configs=[_Config(name=config_key, value=config_value)
-                                 for config_key, config_value in new_topic.topic_configs.items()]
-                    ))
-        if not topics:
-            raise ValueError(f"No valid topics found in new_topics: {new_topics}")
-
-        request = CreateTopicsRequest(
-            topics=topics,
-            timeout_ms=timeout_ms,
-            validate_only=validate_only)
-        def response_errors(r):
-            for topic in r.topics:
-                yield Errors.for_code(topic.error_code)
-        response = self._manager.run(self._send_request_to_controller, request, response_errors, raise_errors)
-        if wait_for_metadata: # implies not validate_only
-            self.wait_for_topics([new_topic.name for new_topic in new_topics])
-        return response
-
-    def wait_for_topics(self, topic_names, timeout_ms=10000):
-        """Block until each of the given topics is ready to use.
-
-        CreateTopicsResponse only confirms that the broker accepted the create
-        request; propagating the new topics into the broker's metadata cache --
-        and electing a leader for every partition -- can lag behind, especially
-        on KRaft clusters. This method polls :meth:`describe_topics` at a fixed
-        interval until every requested topic both:
-
-          - is returned with ``error_code == 0`` (topic exists and is
-            visible in metadata), and
-          - has ``error_code == 0`` and a leader assigned (``leader_id >= 0``)
-            for every partition.
-
-        Useful after :meth:`create_topics` (including implicit creation via
-        ``allow_auto_topic_creation``) or after a delete+recreate sequence.
-
-        Arguments:
-            topic_names ([str]): Topic names to wait for.
-
-        Keyword Arguments:
-            timeout_ms (numeric, optional): Maximum milliseconds to wait.
-                Default: 10000.
-
-        Raises:
-            KafkaTimeoutError: if any topic is still not ready when the
-                deadline expires. The exception message includes the
-                per-topic state from the final ``describe_topics`` call.
-        """
-        if not topic_names:
-            return
-        topic_names = list(topic_names)
-        deadline = time.monotonic() + (timeout_ms / 1000.0)
-        pending = {name: 'not yet queried' for name in topic_names}
-        while True:
-            try:
-                topics = self.describe_topics(topics=topic_names)
-            except Exception as exc:
-                log.debug('describe_topics failed while waiting for topic visibility: %s', exc)
-                topics = []
-            by_name = {t.get('name'): t for t in topics}
-            pending = {}
-            for name in topic_names:
-                reason = self._topic_not_ready_reason(by_name.get(name))
-                if reason is not None:
-                    pending[name] = reason
-            if not pending:
-                return
-            if time.monotonic() >= deadline:
-                raise Errors.KafkaTimeoutError(
-                    'Topics not ready after %sms: %s' % (timeout_ms, pending))
-            time.sleep(0.1)
-
-    @staticmethod
-    def _topic_not_ready_reason(topic_info):
-        """Return a string reason if ``topic_info`` isn't ready, else None."""
-        if topic_info is None:
-            return 'missing from metadata response'
-        error_code = topic_info.get('error_code', 0)
-        if error_code != 0:
-            return Errors.for_code(error_code).__name__
-        partitions = topic_info.get('partitions') or []
-        if not partitions:
-            return 'no partitions reported'
-        bad = []
-        for p in partitions:
-            p_err = p.get('error_code', 0)
-            idx = p.get('partition_index')
-            if p_err != 0:
-                bad.append('p%s=%s' % (idx, Errors.for_code(p_err).__name__))
-                continue
-            if p.get('leader_id', -1) < 0:
-                bad.append('p%s=no leader' % idx)
-        if bad:
-            return ','.join(bad)
-        return None
-
-    def delete_topics(self, topics, timeout_ms=None, raise_errors=True):
-        """Delete topics from the cluster.
-
-        Arguments:
-            topics ([str]): A list of topic name strings or uuid.UUID ids.
-
-        Keyword Arguments:
-            timeout_ms (numeric, optional): Milliseconds to wait for topics to be deleted
-                before the broker returns.
-            raise_errors (bool, optional): Whether to raise errors as exceptions. Default True.
-
-        Returns:
-            Appropriate version of DeleteTopicsResponse class.
-        """
-        timeout_ms = self._validate_timeout(timeout_ms)
-        _Topic = DeleteTopicsRequest.DeleteTopicState
-        request = DeleteTopicsRequest(
-            topics=[_Topic(topic_id=topic) if isinstance(topic, uuid.UUID) else _Topic(name=topic)
-                    for topic in topics],
-            timeout_ms=timeout_ms)
-
-        def response_errors(r):
-            for response in r.responses:
-                yield Errors.for_code(response.error_code)
-        return self._manager.run(self._send_request_to_controller, request, response_errors, raise_errors)
-
-    def _process_acl_operations(self, obj):
-        if obj.get('authorized_operations', None) is not None:
-            obj['authorized_operations'] = list(map(lambda acl: acl.name, valid_acl_operations(obj['authorized_operations'])))
-        return obj
-
-    async def _get_cluster_metadata(self, topics):
-        """topics = [] for no topics, None for all."""
-        request = MetadataRequest(
-            topics=[
-                MetadataRequest.MetadataRequestTopic(name=topic)
-                for topic in topics] if topics is not None else None,
-            allow_auto_topic_creation=False,
-            include_cluster_authorized_operations=True,
-            include_topic_authorized_operations=True,
-        )
-        response = await self._manager.send(request)
-        metadata = response.to_dict()
-        self._process_acl_operations(metadata)
-        for topic in metadata['topics']:
-            self._process_acl_operations(topic)
-        return metadata
-
-    def list_topics(self):
-        """Retrieve a list of all topic names in the cluster.
-
-        Returns:
-            A list of topic name strings.
-        """
-        metadata = self._manager.run(self._get_cluster_metadata, None) # None => request all topics
-        return [t['name'] for t in metadata['topics']]
-
-    def describe_topics(self, topics=None):
-        """Fetch metadata for the specified topics or all topics if None.
-
-        Keyword Arguments:
-            topics ([str], optional) A list of topic names. If None, metadata for all
-                topics is retrieved.
-
-        Returns:
-            A list of dicts describing each topic (including partition info).
-        """
-        metadata = self._manager.run(self._get_cluster_metadata, topics)
-        return metadata['topics']
-
-    def describe_cluster(self):
-        """Fetch cluster-wide metadata such as the list of brokers, the controller ID,
-        and the cluster ID.
-
-        Returns:
-            A dict with cluster-wide metadata, excluding topic details.
-        """
-        metadata = self._manager.run(self._get_cluster_metadata, []) # [] => no topics
-        metadata.pop('topics')  # We have 'describe_topics' for this
-        return metadata
-
-    @staticmethod
-    def _convert_describe_acls_response_to_acls(describe_response):
-        """Convert a DescribeAclsResponse into a list of ACL objects and a KafkaError.
-
-        Arguments:
-            describe_response: The response object from the DescribeAclsRequest.
-
-        Returns:
-            A tuple of (list_of_acl_objects, error) where error is an instance
-                 of KafkaError (NoError if successful).
-        """
-        version = describe_response.API_VERSION
-
-        error = Errors.for_code(describe_response.error_code)
-        acl_list = []
-        for resources in describe_response.resources:
-            if version == 0:
-                resource_type, resource_name, acls = resources
-                resource_pattern_type = ACLResourcePatternType.LITERAL.value
-            elif version <= 1:
-                resource_type, resource_name, resource_pattern_type, acls = resources
-            else:
-                raise NotImplementedError(
-                    "Support for DescribeAcls Response v{} has not yet been added to KafkaAdmin."
-                        .format(version)
-                )
-            for acl in acls:
-                principal, host, operation, permission_type = acl
-                conv_acl = ACL(
-                    principal=principal,
-                    host=host,
-                    operation=ACLOperation(operation),
-                    permission_type=ACLPermissionType(permission_type),
-                    resource_pattern=ResourcePattern(
-                        ResourceType(resource_type),
-                        resource_name,
-                        ACLResourcePatternType(resource_pattern_type)
-                    )
-                )
-                acl_list.append(conv_acl)
-
-        return (acl_list, error,)
-
-    def describe_acls(self, acl_filter):
-        """Describe a set of ACLs
-
-        Used to return a set of ACLs matching the supplied ACLFilter.
-        The cluster must be configured with an authorizer for this to work, or
-        you will get a SecurityDisabledError
-
-        Arguments:
-            acl_filter: an ACLFilter object
-
-        Returns:
-            tuple of a list of matching ACL objects and a KafkaError (NoError if successful)
-        """
-
-        version = self._client.api_version(DescribeAclsRequest, max_version=1)
-        if version == 0:
-            request = DescribeAclsRequest[version](
-                resource_type_filter=acl_filter.resource_pattern.resource_type,
-                resource_name_filter=acl_filter.resource_pattern.resource_name,
-                principal_filter=acl_filter.principal,
-                host_filter=acl_filter.host,
-                operation=acl_filter.operation,
-                permission_type=acl_filter.permission_type
-            )
-        elif version <= 1:
-            request = DescribeAclsRequest[version](
-                resource_type_filter=acl_filter.resource_pattern.resource_type,
-                resource_name_filter=acl_filter.resource_pattern.resource_name,
-                pattern_type_filter=acl_filter.resource_pattern.pattern_type,
-                principal_filter=acl_filter.principal,
-                host_filter=acl_filter.host,
-                operation=acl_filter.operation,
-                permission_type=acl_filter.permission_type
-            )
-        response = self._manager.run(self._manager.send(request))  # pylint: disable=E0606
-        error_type = Errors.for_code(response.error_code)
-        if error_type is not Errors.NoError:
-            # optionally we could retry if error_type.retriable
-            raise error_type(
-                "Request '{}' failed with response '{}'."
-                    .format(request, response))
-
-        return self._convert_describe_acls_response_to_acls(response)
-
-    @staticmethod
-    def _convert_create_acls_resource_request_v0(acl):
-        """Convert an ACL object into the CreateAclsRequest v0 format.
-
-        Arguments:
-            acl: An ACL object with resource pattern and permissions.
-
-        Returns:
-            A tuple: (resource_type, resource_name, principal, host, operation, permission_type).
-        """
-
-        return (
-            acl.resource_pattern.resource_type,
-            acl.resource_pattern.resource_name,
-            acl.principal,
-            acl.host,
-            acl.operation,
-            acl.permission_type
-        )
-
-    @staticmethod
-    def _convert_create_acls_resource_request_v1(acl):
-        """Convert an ACL object into the CreateAclsRequest v1 format.
-
-        Arguments:
-            acl: An ACL object with resource pattern and permissions.
-
-        Returns:
-            A tuple: (resource_type, resource_name, pattern_type, principal, host, operation, permission_type).
-        """
-        return (
-            acl.resource_pattern.resource_type,
-            acl.resource_pattern.resource_name,
-            acl.resource_pattern.pattern_type,
-            acl.principal,
-            acl.host,
-            acl.operation,
-            acl.permission_type
-        )
-
-    @staticmethod
-    def _convert_create_acls_response_to_acls(acls, create_response):
-        """Parse CreateAclsResponse and correlate success/failure with original ACL objects.
-
-        Arguments:
-            acls: A list of ACL objects that were requested for creation.
-            create_response: The broker's CreateAclsResponse object.
-
-        Returns:
-            A dict with:
-                 {
-                   'succeeded': [list of ACL objects successfully created],
-                   'failed': [(acl_object, KafkaError), ...]
-                 }
-        """
-        version = create_response.API_VERSION
-
-        creations_error = []
-        creations_success = []
-        for i, creations in enumerate(create_response.results):
-            if version <= 1:
-                error_code, error_message = creations
-                acl = acls[i]
-                error = Errors.for_code(error_code)
-            else:
-                raise NotImplementedError(
-                    "Support for DescribeAcls Response v{} has not yet been added to KafkaAdmin."
-                        .format(version)
-                )
-
-            if error is Errors.NoError:
-                creations_success.append(acl)
-            else:
-                creations_error.append((acl, error,))
-
-        return {"succeeded": creations_success, "failed": creations_error}
-
-    def create_acls(self, acls):
-        """Create a list of ACLs
-
-        This endpoint only accepts a list of concrete ACL objects, no ACLFilters.
-        Throws TopicAlreadyExistsError if topic is already present.
-
-        Arguments:
-            acls: a list of ACL objects
-
-        Returns:
-            dict of successes and failures
-        """
-
-        for acl in acls:
-            if not isinstance(acl, ACL):
-                raise IllegalArgumentError("acls must contain ACL objects")
-
-        version = self._client.api_version(CreateAclsRequest, max_version=1)
-        if version == 0:
-            request = CreateAclsRequest[version](
-                creations=[self._convert_create_acls_resource_request_v0(acl) for acl in acls]
-            )
-        elif version <= 1:
-            request = CreateAclsRequest[version](
-                creations=[self._convert_create_acls_resource_request_v1(acl) for acl in acls]
-            )
-        response = self._manager.run(self._manager.send(request))  # pylint: disable=E0606
-        return self._convert_create_acls_response_to_acls(acls, response)
-
-    @staticmethod
-    def _convert_delete_acls_resource_request_v0(acl):
-        """Convert an ACLFilter object into the DeleteAclsRequest v0 format.
-
-        Arguments:
-            acl: An ACLFilter object identifying the ACLs to be deleted.
-
-        Returns:
-            A tuple: (resource_type, resource_name, principal, host, operation, permission_type).
-        """
-        return (
-            acl.resource_pattern.resource_type,
-            acl.resource_pattern.resource_name,
-            acl.principal,
-            acl.host,
-            acl.operation,
-            acl.permission_type
-        )
-
-    @staticmethod
-    def _convert_delete_acls_resource_request_v1(acl):
-        """Convert an ACLFilter object into the DeleteAclsRequest v1 format.
-
-        Arguments:
-            acl: An ACLFilter object identifying the ACLs to be deleted.
-
-        Returns:
-            A tuple: (resource_type, resource_name, pattern_type, principal, host, operation, permission_type).
-        """
-        return (
-            acl.resource_pattern.resource_type,
-            acl.resource_pattern.resource_name,
-            acl.resource_pattern.pattern_type,
-            acl.principal,
-            acl.host,
-            acl.operation,
-            acl.permission_type
-        )
-
-    @staticmethod
-    def _convert_delete_acls_response_to_matching_acls(acl_filters, delete_response):
-        """Parse the DeleteAclsResponse and map the results back to each input ACLFilter.
-
-        Arguments:
-            acl_filters: A list of ACLFilter objects that were provided in the request.
-            delete_response: The response from the DeleteAclsRequest.
-
-        Returns:
-            A list of tuples of the form:
-                 (acl_filter, [(matching_acl, KafkaError), ...], filter_level_error).
-        """
-        version = delete_response.API_VERSION
-        filter_result_list = []
-        for i, filter_responses in enumerate(delete_response.filter_results):
-            filter_error_code, filter_error_message, matching_acls = filter_responses
-            filter_error = Errors.for_code(filter_error_code)
-            acl_result_list = []
-            for acl in matching_acls:
-                if version == 0:
-                    error_code, error_message, resource_type, resource_name, principal, host, operation, permission_type = acl
-                    resource_pattern_type = ACLResourcePatternType.LITERAL.value
-                elif version == 1:
-                    error_code, error_message, resource_type, resource_name, resource_pattern_type, principal, host, operation, permission_type = acl
-                else:
-                    raise NotImplementedError(
-                        "Support for DescribeAcls Response v{} has not yet been added to KafkaAdmin."
-                            .format(version)
-                    )
-                acl_error = Errors.for_code(error_code)
-                conv_acl = ACL(
-                    principal=principal,
-                    host=host,
-                    operation=ACLOperation(operation),
-                    permission_type=ACLPermissionType(permission_type),
-                    resource_pattern=ResourcePattern(
-                        ResourceType(resource_type),
-                        resource_name,
-                        ACLResourcePatternType(resource_pattern_type)
-                    )
-                )
-                acl_result_list.append((conv_acl, acl_error,))
-            filter_result_list.append((acl_filters[i], acl_result_list, filter_error,))
-        return filter_result_list
-
-    def delete_acls(self, acl_filters):
-        """Delete a set of ACLs
-
-        Deletes all ACLs matching the list of input ACLFilter
-
-        Arguments:
-            acl_filters: a list of ACLFilter
-
-        Returns:
-            a list of 3-tuples corresponding to the list of input filters.
-                 The tuples hold (the input ACLFilter, list of affected ACLs, KafkaError instance)
-        """
-
-        for acl in acl_filters:
-            if not isinstance(acl, ACLFilter):
-                raise IllegalArgumentError("acl_filters must contain ACLFilter type objects")
-
-        version = self._client.api_version(DeleteAclsRequest, max_version=1)
-
-        if version == 0:
-            request = DeleteAclsRequest[version](
-                filters=[self._convert_delete_acls_resource_request_v0(acl) for acl in acl_filters]
-            )
-        elif version <= 1:
-            request = DeleteAclsRequest[version](
-                filters=[self._convert_delete_acls_resource_request_v1(acl) for acl in acl_filters]
-            )
-        response = self._manager.run(self._manager.send(request))  # pylint: disable=E0606
-        return self._convert_delete_acls_response_to_matching_acls(acl_filters, response)
-
-    @staticmethod
-    def _convert_describe_config_resource_request(config_resource):
-        """Convert a ConfigResource into the format required by DescribeConfigsRequest.
-
-        Arguments:
-            config_resource: A ConfigResource with resource_type, name, and optional config keys.
-
-        Returns:
-            A tuple: (resource_type, resource_name, [list_of_config_keys] or None).
-        """
-        return (
-            config_resource.resource_type,
-            config_resource.name,
-            [
-                config_key for config_key, config_value in config_resource.configs.items()
-            ] if config_resource.configs else None
-        )
-
-    async def _async_describe_configs(self, config_resources, include_synonyms=False):
-        # Break up requests by type - a broker config request must be sent to the specific broker.
-        # All other (currently just topic resources) can be sent to any broker.
-        broker_resources = []
-        topic_resources = []
-
-        for config_resource in config_resources:
-            if config_resource.resource_type == ConfigResourceType.BROKER:
-                broker_resources.append(self._convert_describe_config_resource_request(config_resource))
-            else:
-                topic_resources.append(self._convert_describe_config_resource_request(config_resource))
-
-        version = self._client.api_version(DescribeConfigsRequest, max_version=2)
-        if include_synonyms and version == 0:
-            raise IncompatibleBrokerVersion(
-                "include_synonyms requires DescribeConfigsRequest >= v1, which is not supported by Kafka {}."
-                    .format(self._manager.broker_version))
-
-        results = []
-        for broker_resource in broker_resources:
-            try:
-                broker_id = int(broker_resource[1])
-            except ValueError:
-                raise ValueError("Broker resource names must be an integer or a string represented integer")
-            if version == 0:
-                request = DescribeConfigsRequest[version](resources=[broker_resource])
-            else:
-                request = DescribeConfigsRequest[version](
-                    resources=[broker_resource],
-                    include_synonyms=include_synonyms)
-            results.append(await self._manager.send(request, node_id=broker_id))
-
-        if topic_resources:
-            if version == 0:
-                request = DescribeConfigsRequest[version](resources=topic_resources)
-            else:
-                request = DescribeConfigsRequest[version](resources=topic_resources, include_synonyms=include_synonyms)
-            results.append(await self._manager.send(request))
-
-        return results
-
-    def describe_configs(self, config_resources, include_synonyms=False):
-        """Fetch configuration parameters for one or more Kafka resources.
-
-        Arguments:
-            config_resources: An list of ConfigResource objects.
-                Any keys in ConfigResource.configs dict will be used to filter the
-                result. Setting the configs dict to None will get all values. An
-                empty dict will get zero values (as per Kafka protocol).
-
-        Keyword Arguments:
-            include_synonyms (bool, optional): If True, return synonyms in response. Not
-                supported by all versions. Default: False.
-
-        Returns:
-            List of DescribeConfigsResponses.
-        """
-        return self._manager.run(self._async_describe_configs, config_resources, include_synonyms)
-
-    @staticmethod
-    def _convert_alter_config_resource_request(config_resource):
-        """Convert a ConfigResource into the format required by AlterConfigsRequest.
-
-        Arguments:
-            config_resource: A ConfigResource with resource_type, name, and config (key, value) pairs.
-
-        Returns:
-            A tuple: (resource_type, resource_name, [(config_key, config_value), ...]).
-        """
-        return (
-            config_resource.resource_type,
-            config_resource.name,
-            [
-                (config_key, config_value) for config_key, config_value in config_resource.configs.items()
-            ]
-        )
-
-    async def _async_alter_configs(self, config_resources):
-        version = self._client.api_version(AlterConfigsRequest, max_version=1)
-        request = AlterConfigsRequest[version](
-            resources=[self._convert_alter_config_resource_request(config_resource) for config_resource in config_resources]
-        )
-        # TODO the Java client has the note:
-        # // We must make a separate AlterConfigs request for every BROKER resource we want to alter
-        # // and send the request to that specific broker. Other resources are grouped together into
-        # // a single request that may be sent to any broker.
-        #
-        # So this is currently broken as it always sends to the least_loaded_node()
-        return await self._manager.send(request)
-
-    def alter_configs(self, config_resources):
-        """Alter configuration parameters of one or more Kafka resources.
-
-        Warning:
-            This is currently broken for BROKER resources because those must be
-            sent to that specific broker, versus this always picks the
-            least-loaded node. See the comment in the source code for details.
-            We would happily accept a PR fixing this.
-
-        Arguments:
-            config_resources: A list of ConfigResource objects.
-
-        Returns:
-            Appropriate version of AlterConfigsResponse class.
-        """
-        return self._manager.run(self._async_alter_configs, config_resources)
-
-    # alter replica logs dir protocol not yet implemented
-    # Note: have to lookup the broker with the replica assignment and send the request to that broker
-
-    # describe log dirs protocol not yet implemented
-    # Note: have to lookup the broker with the replica assignment and send the request to that broker
-
-    def create_partitions(self, topic_partitions, timeout_ms=None, validate_only=False, raise_errors=True):
-        """Create additional partitions for an existing topic.
-
-        Arguments:
-            topic_partitions: A dict of topic name strings to total partition count (int),
-                or a dict of {topic_name: {count: int, assignments: [[broker_ids]]}}
-                if manual assignment is desired.
-                dict of {topic_name: NewPartition} is deprecated.
-
-        Keyword Arguments:
-            timeout_ms (numeric, optional): Milliseconds to wait for new partitions to be
-                created before the broker returns.
-            validate_only (bool, optional): If True, don't actually create new partitions.
-                Default: False
-            raise_errors (bool, optional): Whether to raise errors as exceptions. Default True.
-
-        Returns:
-            Appropriate version of CreatePartitionsResponse class.
-        """
-        timeout_ms = self._validate_timeout(timeout_ms)
-        _Topic = CreatePartitionsRequest.CreatePartitionsTopic
-        _Assignment = CreatePartitionsRequest.CreatePartitionsTopic.CreatePartitionsAssignment
-        topics = []
-        for topic, count in topic_partitions.items():
-            if isinstance(count, int):
-                topics.append(_Topic(name=topic, count=count))
-            elif isinstance(count, dict):
-                topics.append(
-                    _Topic(
-                        name=topic,
-                        count=count['count'],
-                        assignments=[_Assignment(broker_ids=broker_ids)
-                                     for broker_ids in count['assignments']]))
-
-            else:
-                topics.append(
-                    _Topic(
-                        name=topic,
-                        count=count.total_count,
-                        assignments=[_Assignment(broker_ids=broker_ids)
-                                     for broker_ids in count.new_assignments]))
-        request = CreatePartitionsRequest(
-            topics=topics,
-            timeout_ms=timeout_ms,
-            validate_only=validate_only)
-
-        def response_errors(r):
-            for result in r.results:
-                yield Errors.for_code(result.error_code)
-        return self._manager.run(self._send_request_to_controller, request, response_errors, raise_errors)
-
-    async def _async_get_leader_for_partitions(self, partitions):
-        """Finds ID of the leader node for every given topic partition."""
-        partitions = set(partitions)
-        topics = set(tp.topic for tp in partitions)
-
-        metadata = await self._get_cluster_metadata(topics)
-
-        leader2partitions = defaultdict(list)
-        valid_partitions = set()
-        for topic in metadata.get("topics", ()):
-            for partition in topic.get("partitions", ()):
-                t2p = TopicPartition(topic=topic["name"], partition=partition["partition_index"])
-                if t2p in partitions:
-                    leader2partitions[partition["leader_id"]].append(t2p)
-                    valid_partitions.add(t2p)
-
-        if len(partitions) != len(valid_partitions):
-            unknown = set(partitions) - valid_partitions
-            raise UnknownTopicOrPartitionError(
-                "The following partitions are not known: %s"
-                % ", ".join(str(x) for x in unknown)
-            )
-
-        return leader2partitions
-
-    async def _async_delete_records(self, records_to_delete, timeout_ms=None, partition_leader_id=None):
-        timeout_ms = self._validate_timeout(timeout_ms)
-        version = self._client.api_version(DeleteRecordsRequest, max_version=0)
-
-        if partition_leader_id is None:
-            leader2partitions = await self._async_get_leader_for_partitions(set(records_to_delete))
-        else:
-            leader2partitions = {partition_leader_id: set(records_to_delete)}
-
-        responses = []
-        for leader, partitions in leader2partitions.items():
-            topic2partitions = defaultdict(list)
-            for partition in partitions:
-                topic2partitions[partition.topic].append(partition)
-
-            request = DeleteRecordsRequest[version](
-                topics=[
-                    (topic, [(tp.partition, records_to_delete[tp]) for tp in partitions])
-                    for topic, partitions in topic2partitions.items()
-                ],
-                timeout_ms=timeout_ms
-            )
-            response = await self._manager.send(request, node_id=leader)
-            responses.append(response.to_dict())
-
-        partition2result = {}
-        partition2error = {}
-        for response in responses:
-            for topic in response["topics"]:
-                for partition in topic["partitions"]:
-                    tp = TopicPartition(topic["name"], partition["partition_index"])
-                    partition2result[tp] = partition
-                    if partition["error_code"] != 0:
-                        partition2error[tp] = partition["error_code"]
-
-        if partition2error:
-            if len(partition2error) == 1:
-                key, error = next(iter(partition2error.items()))
-                raise Errors.for_code(error)(
-                    "Error deleting records from topic %s partition %s" % (key.topic, key.partition)
-                )
-            else:
-                raise Errors.BrokerResponseError(
-                    "The following errors occured when trying to delete records: " +
-                    ", ".join(
-                        "%s(partition=%d): %s" %
-                        (partition.topic, partition.partition, Errors.for_code(error).__name__)
-                        for partition, error in partition2error.items()
-                    )
-                )
-
-        return partition2result
-
-    def delete_records(self, records_to_delete, timeout_ms=None, partition_leader_id=None):
-        """Delete records whose offset is smaller than the given offset of the corresponding partition.
-
-        Arguments:
-            records_to_delete ({TopicPartition: int}): The earliest available offsets for the
-                given partitions.
-
-        Keyword Arguments:
-            timeout_ms (numeric, optional): Timeout in milliseconds, if None (default), will be read from
-                config.
-            partition_leader_id (node_id / int, optional): If specified, all deletion requests will be sent to
-                this node. No check is performed verifying that this is indeed the leader for all
-                listed partitions: use with caution.
-
-        Returns:
-            dict {topicPartition -> metadata}, where metadata is returned by the broker.
-            See DeleteRecordsResponse for possible fields. error_code for all partitions is
-            guaranteed to be zero, otherwise an exception is raised.
-        """
-        return self._manager.run(self._async_delete_records, records_to_delete, timeout_ms, partition_leader_id)
-
-    # create delegation token protocol not yet implemented
-    # Note: send the request to the least_loaded_node()
-
-    # renew delegation token protocol not yet implemented
-    # Note: send the request to the least_loaded_node()
-
-    # expire delegation_token protocol not yet implemented
-    # Note: send the request to the least_loaded_node()
-
-    # describe delegation_token protocol not yet implemented
-    # Note: send the request to the least_loaded_node()
-
-    def _describe_consumer_groups_request(self, group_id):
-        """Send a DescribeGroupsRequest to the group's coordinator.
-
-        Arguments:
-            group_id: The group name as a string
-
-        Returns:
-            DescribeGroupsRequest object
-        """
-        version = self._client.api_version(DescribeGroupsRequest, max_version=3)
-        if version <= 2:
-            # Note: KAFKA-6788 A potential optimization is to group the
-            # request per coordinator and send one request with a list of
-            # all consumer groups. Java still hasn't implemented this
-            # because the error checking is hard to get right when some
-            # groups error and others don't.
-            request = DescribeGroupsRequest[version](groups=(group_id,))
-        else:
-            request = DescribeGroupsRequest[version](
-                groups=(group_id,),
-                include_authorized_operations=True
-            )
-        return request
-
-    def _describe_consumer_groups_process_response(self, response):
-        """Process a DescribeGroupsResponse into a group description."""
-        if response.API_VERSION > 3:
-            raise NotImplementedError(
-                "Support for DescribeGroupsResponse_v{} has not yet been added to KafkaAdminClient."
-                .format(response.API_VERSION))
-
-        assert len(response.groups) == 1
-        for response_name, response_field in response.fields.items():
-            if response_name == 'groups':
-                described_group = getattr(response, response_name)[0]
-                described_group_information_list = []
-                protocol_type_is_consumer = False
-                for group_information_name, group_information_field in response_field.fields.items():
-                    if not group_information_field.for_version_q(response.API_VERSION):
-                        continue
-                    described_group_information = getattr(described_group, group_information_name)
-                    if group_information_name == 'protocol_type':
-                        protocol_type = described_group_information
-                        protocol_type_is_consumer = (protocol_type == ConsumerProtocolType or not protocol_type)
-                    if group_information_name == 'members':
-                        member_information_list = []
-                        for member in described_group_information:
-                            member_information = []
-                            for attr_name, attr_field in group_information_field.fields.items():
-                                if not attr_field.for_version_q(response.API_VERSION):
-                                    continue
-                                attr_val = getattr(member, attr_name)
-                                if protocol_type_is_consumer:
-                                    if attr_name == 'member_metadata' and attr_val:
-                                        member_information.append(ConsumerProtocolSubscription.decode(attr_val))
-                                    elif attr_name == 'member_assignment' and attr_val:
-                                        member_information.append(ConsumerProtocolAssignment.decode(attr_val))
-                                    else:
-                                        member_information.append(attr_val)
-                            member_info_tuple = MemberInformation._make(member_information)
-                            member_information_list.append(member_info_tuple)
-                        described_group_information_list.append(member_information_list)
-                    else:
-                        described_group_information_list.append(described_group_information)
-                # Version 3 of the DescribeGroups API introduced the "authorized_operations" field.
-                if response.API_VERSION >= 3:
-                    described_group_information_list[-1] = list(map(lambda acl: acl.name, valid_acl_operations(described_group_information_list[-1])))
-                else:
-                    # TODO: Fix GroupInformation defaults
-                    described_group_information_list.append([])
-                group_description = GroupInformation._make(described_group_information_list)
-                error_code = group_description.error_code
-                error_type = Errors.for_code(error_code)
-                # Java has the note: KAFKA-6789, we can retry based on the error code
-                if error_type is not Errors.NoError:
-                    raise error_type(
-                        "DescribeGroupsResponse failed with response '{}'."
-                        .format(response))
-                return group_description
-        assert False, "DescribeGroupsResponse parsing failed"
-
-    async def _async_describe_consumer_groups(self, group_ids, group_coordinator_id=None):
-        results = []
-        for group_id in group_ids:
-            coordinator_id = group_coordinator_id or await self._find_coordinator_id(group_id)
-            request = self._describe_consumer_groups_request(group_id)
-            response = await self._manager.send(request, node_id=coordinator_id)
-            results.append(self._describe_consumer_groups_process_response(response))
-        return results
-
-    def describe_consumer_groups(self, group_ids, group_coordinator_id=None, include_authorized_operations=False):
-        """Describe a set of consumer groups.
-
-        Any errors are immediately raised.
-
-        Arguments:
-            group_ids: A list of consumer group IDs. These are typically the
-                group names as strings.
-
-        Keyword Arguments:
-            group_coordinator_id (int, optional): The node_id of the groups' coordinator
-                broker. If set to None, it will query the cluster for each group to
-                find that group's coordinator. Explicitly specifying this can be
-                useful for avoiding extra network round trips if you already know
-                the group coordinator. This is only useful when all the group_ids
-                have the same coordinator, otherwise it will error. Default: None.
-
-        Returns:
-            A list of group descriptions. For now the group descriptions
-            are the raw results from the DescribeGroupsResponse. Long-term, we
-            plan to change this to return namedtuples as well as decoding the
-            partition assignments.
-        """
-        return self._manager.run(self._async_describe_consumer_groups, group_ids, group_coordinator_id)
-
-    def _list_consumer_groups_request(self):
-        """Send a ListGroupsRequest to a broker.
-
-        Returns:
-            ListGroupsRequest object
-        """
-        version = self._client.api_version(ListGroupsRequest, max_version=2)
-        return ListGroupsRequest[version]()
-
-    def _list_consumer_groups_process_response(self, response):
-        """Process a ListGroupsResponse into a list of groups."""
-        if response.API_VERSION <= 2:
-            error_type = Errors.for_code(response.error_code)
-            if error_type is not Errors.NoError:
-                raise error_type(
-                    "ListGroupsRequest failed with response '{}'."
-                    .format(response))
-        else:
-            raise NotImplementedError(
-                "Support for ListGroupsResponse_v{} has not yet been added to KafkaAdminClient."
-                .format(response.API_VERSION))
-        return [(group.group_id, group.protocol_type) for group in response.groups]
-
-    async def _async_list_consumer_groups(self, broker_ids=None):
-        if broker_ids is None:
-            broker_ids = [broker.node_id for broker in self._manager.cluster.brokers()]
-        # Use a set to prevent duplicates if a group coordinator migrates
-        # between brokers mid-query.
-        consumer_groups = set()
-        for broker_id in broker_ids:
-            request = self._list_consumer_groups_request()
-            response = await self._manager.send(request, node_id=broker_id)
-            consumer_groups.update(self._list_consumer_groups_process_response(response))
-        return list(consumer_groups)
-
-    def list_consumer_groups(self, broker_ids=None):
-        """List all consumer groups known to the cluster.
-
-        This returns a list of Consumer Group tuples. The tuples are
-        composed of the consumer group name and the consumer group protocol
-        type.
-
-        Only consumer groups that store their offsets in Kafka are returned.
-        The protocol type will be an empty string for groups created using
-        Kafka < 0.9 APIs because, although they store their offsets in Kafka,
-        they don't use Kafka for group coordination. For groups created using
-        Kafka >= 0.9, the protocol type will typically be "consumer".
-
-        As soon as any error is encountered, it is immediately raised.
-
-        Keyword Arguments:
-            broker_ids ([int], optional): A list of broker node_ids to query for consumer
-                groups. If set to None, will query all brokers in the cluster.
-                Explicitly specifying broker(s) can be useful for determining which
-                consumer groups are coordinated by those broker(s). Default: None
-
-        Returns:
-            list: List of tuples of Consumer Groups.
-
-        Raises:
-            CoordinatorNotAvailableError: The coordinator is not
-                available, so cannot process requests.
-            CoordinatorLoadInProgressError: The coordinator is loading and
-                hence can't process requests.
-        """
-        return self._manager.run(self._async_list_consumer_groups, broker_ids)
-
-    def _list_consumer_group_offsets_request(self, group_id, partitions=None):
-        """Send an OffsetFetchRequest to a broker.
-
-        Arguments:
-            group_id (str): The consumer group id name for which to fetch offsets.
-
-        Keyword Arguments:
-            partitions: A list of TopicPartitions for which to fetch
-                offsets. On brokers >= 0.10.2, this can be set to None to fetch all
-                known offsets for the consumer group. Default: None.
-
-        Returns:
-            OffsetFetchRequest object
-        """
-        version = self._client.api_version(OffsetFetchRequest, max_version=5)
-        if partitions is None:
-            if version <= 1:
-                raise ValueError(
-                    """OffsetFetchRequest_v{} requires specifying the
-                    partitions for which to fetch offsets. Omitting the
-                    partitions is only supported on brokers >= 0.10.2.
-                    For details, see KIP-88.""".format(version))
-            topics_partitions = None
-        else:
-            # transform from [TopicPartition("t1", 1), TopicPartition("t1", 2)] to [("t1", [1, 2])]
-            topics_partitions_dict = defaultdict(set)
-            for topic, partition in partitions:
-                topics_partitions_dict[topic].add(partition)
-            topics_partitions = list(topics_partitions_dict.items())
-        return OffsetFetchRequest[version](group_id, topics_partitions)
-
-    def _list_consumer_group_offsets_process_response(self, response):
-        """Process an OffsetFetchResponse.
-
-        Arguments:
-            response: an OffsetFetchResponse.
-
-        Returns:
-            A dictionary composed of TopicPartition keys and
-            OffsetAndMetadata values.
-        """
-        if response.API_VERSION <= 5:
-
-            # OffsetFetchResponse_v1 lacks a top-level error_code
-            if response.API_VERSION > 1:
-                error_type = Errors.for_code(response.error_code)
-                if error_type is not Errors.NoError:
-                    # optionally we could retry if error_type.retriable
-                    raise error_type(
-                        "OffsetFetchResponse failed with response '{}'."
-                        .format(response))
-
-            # transform response into a dictionary with TopicPartition keys and
-            # OffsetAndMetadata values--this is what the Java AdminClient returns
-            offsets = {}
-            for topic, partitions in response.topics:
-                for partition_data in partitions:
-                    if response.API_VERSION <= 4:
-                        partition, offset, metadata, error_code = partition_data
-                        leader_epoch = -1
-                    else:
-                        partition, offset, leader_epoch, metadata, error_code = partition_data
-                    error_type = Errors.for_code(error_code)
-                    if error_type is not Errors.NoError:
-                        raise error_type(
-                            "Unable to fetch consumer group offsets for topic {}, partition {}"
-                            .format(topic, partition))
-                    offsets[TopicPartition(topic, partition)] = OffsetAndMetadata(offset, metadata, leader_epoch)
-        else:
-            raise NotImplementedError(
-                "Support for OffsetFetchResponse_v{} has not yet been added to KafkaAdminClient."
-                .format(response.API_VERSION))
-        return offsets
-
-    async def _async_list_consumer_group_offsets(self, group_id, group_coordinator_id=None, partitions=None):
-        if group_coordinator_id is None:
-            group_coordinator_id = await self._find_coordinator_id(group_id)
-        request = self._list_consumer_group_offsets_request(group_id, partitions)
-        response = await self._manager.send(request, node_id=group_coordinator_id)
-        return self._list_consumer_group_offsets_process_response(response)
-
-    def list_consumer_group_offsets(self, group_id, group_coordinator_id=None,
-                                    partitions=None):
-        """Fetch Consumer Offsets for a single consumer group.
-
-        Note:
-        This does not verify that the group_id or partitions actually exist
-        in the cluster.
-
-        As soon as any error is encountered, it is immediately raised.
-
-        Arguments:
-            group_id (str): The consumer group id name for which to fetch offsets.
-
-        Keyword Arguments:
-            group_coordinator_id (int, optional): The node_id of the group's coordinator
-                broker. If set to None, will query the cluster to find the group
-                coordinator. Explicitly specifying this can be useful to prevent
-                that extra network round trip if you already know the group
-                coordinator. Default: None.
-            partitions: A list of TopicPartitions for which to fetch
-                offsets. On brokers >= 0.10.2, this can be set to None to fetch all
-                known offsets for the consumer group. Default: None.
-
-        Returns:
-            dictionary: A dictionary with TopicPartition keys and
-            OffsetAndMetadata values. Partitions that are not specified and for
-            which the group_id does not have a recorded offset are omitted. An
-            offset value of `-1` indicates the group_id has no offset for that
-            TopicPartition. A `-1` can only happen for partitions that are
-            explicitly specified.
-        """
-        return self._manager.run(self._async_list_consumer_group_offsets, group_id, group_coordinator_id, partitions)
-
-    async def _async_delete_consumer_groups(self, group_ids, group_coordinator_id=None):
-        coordinators_groups = defaultdict(list)
-        if group_coordinator_id is not None:
-            coordinators_groups[group_coordinator_id] = group_ids
-        else:
-            for group_id in group_ids:
-                coordinator_id = await self._find_coordinator_id(group_id)
-                coordinators_groups[coordinator_id].append(group_id)
-
-        results = []
-        for coordinator_id, coordinator_group_ids in coordinators_groups.items():
-            request = self._delete_consumer_groups_request(coordinator_group_ids)
-            response = await self._manager.send(request, node_id=coordinator_id)
-            results.extend(self._convert_delete_groups_response(response))
-        return results
-
-    def delete_consumer_groups(self, group_ids, group_coordinator_id=None):
-        """Delete Consumer Group Offsets for given consumer groups.
-
-        Note:
-        This does not verify that the group ids actually exist and
-        group_coordinator_id is the correct coordinator for all these groups.
-
-        The result needs checking for potential errors.
-
-        Arguments:
-            group_ids ([str]): The consumer group ids of the groups which are to be deleted.
-
-        Keyword Arguments:
-            group_coordinator_id (int, optional): The node_id of the broker which is
-                the coordinator for all the groups. Use only if all groups are coordinated
-                by the same broker. If set to None, will query the cluster to find the coordinator
-                for every single group. Explicitly specifying this can be useful to prevent
-                that extra network round trips if you already know the group coordinator.
-                Default: None.
-
-        Returns:
-            A list of tuples (group_id, KafkaError)
-        """
-        return self._manager.run(self._async_delete_consumer_groups, group_ids, group_coordinator_id)
-
-    def _convert_delete_groups_response(self, response):
-        """Parse the DeleteGroupsResponse, mapping group IDs to their respective errors.
-
-        Arguments:
-            response: A DeleteGroupsResponse object from the broker.
-
-        Returns:
-            A list of (group_id, KafkaError) for each deleted group.
-        """
-        if response.API_VERSION <= 1:
-            results = []
-            for group_id, error_code in response.results:
-                results.append((group_id, Errors.for_code(error_code)))
-            return results
-        else:
-            raise NotImplementedError(
-                "Support for DeleteGroupsResponse_v{} has not yet been added to KafkaAdminClient."
-                    .format(response.API_VERSION))
-
-    def _delete_consumer_groups_request(self, group_ids):
-        """Build a DeleteGroupsRequest to send to a broker (the group coordinator).
-
-        Arguments:
-            group_ids ([str]): A list of consumer group IDs to be deleted.
-
-        Returns:
-            A DeleteGroupsRequest object.
-        """
-        version = self._client.api_version(DeleteGroupsRequest, max_version=1)
-        return DeleteGroupsRequest[version](group_ids)
-
-    @staticmethod
-    def _convert_topic_partitions(topic_partitions):
-        return [
-            (
-                topic,
-                partition_ids
-            )
-            for topic, partition_ids in topic_partitions.items()
-        ]
-
-    def _get_all_topic_partitions(self):
-        return [
-            (
-                topic,
-                [partition_info.partition for partition_info in self._client.cluster._partitions[topic].values()]
-            )
-            for topic in self._client.cluster.topics()
-        ]
-
-    def _get_topic_partitions(self, topic_partitions):
-        if topic_partitions is None:
-            return self._get_all_topic_partitions()
-        return self._convert_topic_partitions(topic_partitions)
-
-    def perform_leader_election(self, election_type, topic_partitions=None, timeout_ms=None, raise_errors=True):
-        """Perform leader election on the topic partitions.
-
-        Arguments:
-            election_type: Type of election to attempt. 0 for Preferred, 1 for Unclean
-
-        Keyword Arguments:
-            topic_partitions (dict): A map of topic name strings to partition ids list.
-                By default, will run on all topic partitions
-            timeout_ms (num, optional): Milliseconds to wait for the leader election process to complete
-                before the broker returns.
-            raise_errors (bool, optional): True/False whether to raise errors as exceptions. Default True.
-
-        Returns:
-            Appropriate version of ElectLeadersResponse class.
-        """
-        timeout_ms = self._validate_timeout(timeout_ms)
-        request = ElectLeadersRequest(
-            election_type=ElectionType(election_type),
-            topic_partitions=self._get_topic_partitions(topic_partitions),
-            timeout_ms=timeout_ms)
-        def response_errors(r):
-            if r.API_VERSION >= 1:
-                yield Errors.for_code(r.error_code)
-            for result in r.replica_election_results:
-                for partition in result.partition_result:
-                    yield Errors.for_code(partition.error_code)
-        ignore_errors = (Errors.ElectionNotNeededError,)
-        return self._manager.run(self._send_request_to_controller, request, response_errors, raise_errors, ignore_errors)
-
-    async def _async_describe_log_dirs(self):
-        version = self._client.api_version(DescribeLogDirsRequest, max_version=0)
-        return await self._manager.send(DescribeLogDirsRequest[version]())
-
-    def describe_log_dirs(self):
-        """Send a DescribeLogDirsRequest request to a broker.
-
-        Returns:
-            DescribeLogDirsResponse object
-        """
-        return self._manager.run(self._async_describe_log_dirs)
